feat: 添加多种加密算法，添加VERSION信息，添加自动创建同名用户

+++ scripts/build.sh
@@ -1,9 +1,19 @@
 #!/bin/bash
+
+# 检查是否提供了至少一个参数
+if [ $# -eq 0 ]; then
+    echo "请提供当前版本参数!"
+    echo "使用方法: $0 <VERSION>"
+    exit 1
+fi
+
+
 rm -rf dist
+VERSION=$1
 
 go env -w GOOS=linux
 go env -w CGO_ENABLED=0
-go build -ldflags "-s -w" -o dist/linux/bin/ldapadm ./main.go
+go build -ldflags "-s -w -X ldapadm/Version.VERSION=${VERSION}" -o dist/linux/bin/ldapadm ./main.go
 echo "Linux build success!"
 
 mkdir dist/linux/etc
